myapp/auth/jwt.py
from datetime import datetime, timedelta, UTC
from jose import jwt
import secrets

# JWT設定
SECRET_KEY = secrets.token_urlsafe(32)
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# --------------------------
# JWT作成
# --------------------------
def create_access_token(data: dict)->str:
    """
    JWTを作成する
    """
    to_encode = data.copy()
    expire = datetime.now(UTC) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
