requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[argon2]
pytest
httpx